moved User.jsx to admin
