+++ src/components/admin/User.jsx
@@ -1,11 +1,11 @@
 import React, {useEffect, useState} from 'react';
 import { useLocation } from 'react-router-dom';
-import UserDisplay from './displayData/UserDisplay';
-import OrderHistory from './displayData/OrderHistory';
+import UserDisplay from '../displayData/UserDisplay';
+import OrderHistory from '../displayData/OrderHistory';
 import axios from 'axios';
-import useAdminCheck from "./admin/useAdminCheck";
-import LoadingAndError from "./LoadingAndError"
-import {API_URL} from '../config/config.js';
+import useAdminCheck from "./useAdminCheck";
+import LoadingAndError from "../LoadingAndError"
+import {API_URL} from '../../config/config.js';
 
 const User = () => {
     useAdminCheck();
@@ -13,7 +13,7 @@ const User = () => {
     const [orders, setOrders] = useState([]);
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState('');
-    const location = useLocation();  // Используем useLocation вместо location
+    const location = useLocation();
     const queryParams = new URLSearchParams(location.search);
     const userId = queryParams.get('id');
